+++ theme/__init__.py
@@ -1,1 +1,2 @@
 from .Theme import *
+from .Rule import *
