+++ theme/constants.py
@@ -1,0 +1,21 @@
+
+theme = 'Theme'
+
+quiet = 'Quiet Theme'
+engaged = 'Engaged Theme'
+romantic = 'Romantic Theme'
+cozy = 'Cozy'
+reading = 'Concentrated'
+sleep = 'Sleep'
+
+image = 'Image'
+light = 'Light'
+music = 'Music'
+seat_angle = 'Seat Angle'
+reading_light = 'Reading Light'
+entertainment_recommendation = 'Entertainment Recommendation'
+
+turbulance = 'Turbulance'
+ambiant_light = 'Ambiant Light'
+safety_belt_warning = 'Safety Belt Warning'
+entertainment_pause = 'Entertainment Pause'
