+++ user/user_database.json
@@ -1,1 +1,1 @@
-{"ken": {"birthday": "_", "preference": {"global_theme": null, "personal_theme": "Cozy", "disabled_default_reactions": {"TurbulanceRulePersonal": true}}, "history": ["2019-08-26 15:03:29.586340 Set preference: personal_theme to Cozy.", "2019-08-26 15:03:38.048259 Manual Override: TurbulanceRulePersonal is disabled. Context: Turbulance=85."]}}
+{"ken": {"birthday": "_", "preference": {"global_theme": null, "personal_theme": "Cozy", "disabled_default_reactions": {}}, "history": ["2019-08-26 15:03:29.586340 Set preference: personal_theme to Cozy.", "2019-08-26 15:03:38.048259 Manual Override: TurbulanceRulePersonal is disabled. Context: Turbulance=85."]}, "1": {"birthday": "bd", "preference": {"global_theme": null, "personal_theme": null, "disabled_default_reactions": {"TurbulanceRulePersonal": true}}, "history": []}}
